Updated parseProperty paths to use structure definitions.
