--- src/data/nonMandatoryElements.ts
@@ -1,3543 +1,0 @@
-export const mandatoryElements: Record<string, string[]> = {
-  Account: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'name',
-    'subject',
-    'servicePeriod',
-    'coverage',
-    'owner',
-    'description',
-    'guarantor',
-    'partOf'
-  ],
-  ActivityDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'library',
-    'kind',
-    'profile',
-    'code',
-    'intent',
-    'priority',
-    'doNotPerform',
-    'timing[x]',
-    'location',
-    'participant',
-    'product[x]',
-    'quantity',
-    'dosage',
-    'bodySite',
-    'specimenRequirement',
-    'observationRequirement',
-    'observationResultRequirement',
-    'transform',
-    'dynamicValue'
-  ],
-  AdverseEvent: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'category',
-    'event',
-    'encounter',
-    'date',
-    'detected',
-    'recordedDate',
-    'resultingCondition',
-    'location',
-    'seriousness',
-    'severity',
-    'outcome',
-    'recorder',
-    'contributor',
-    'suspectEntity',
-    'subjectMedicalHistory',
-    'referenceDocument',
-    'study'
-  ],
-  AllergyIntolerance: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'clinicalStatus',
-    'verificationStatus',
-    'type',
-    'category',
-    'criticality',
-    'code',
-    'encounter',
-    'onset[x]',
-    'recordedDate',
-    'recorder',
-    'asserter',
-    'lastOccurrence',
-    'note',
-    'reaction'
-  ],
-  Appointment: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'cancelationReason',
-    'serviceCategory',
-    'serviceType',
-    'specialty',
-    'appointmentType',
-    'reasonCode',
-    'reasonReference',
-    'priority',
-    'description',
-    'supportingInformation',
-    'start',
-    'end',
-    'minutesDuration',
-    'slot',
-    'created',
-    'comment',
-    'patientInstruction',
-    'basedOn',
-    'requestedPeriod'
-  ],
-  AppointmentResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'start',
-    'end',
-    'participantType',
-    'actor',
-    'comment'
-  ],
-  AuditEvent: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'subtype',
-    'action',
-    'period',
-    'outcome',
-    'outcomeDesc',
-    'purposeOfEvent',
-    'entity'
-  ],
-  Basic: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subject',
-    'created',
-    'author'
-  ],
-  Binary: ['id', 'meta', 'implicitRules', 'language', 'securityContext', 'data'],
-  BiologicallyDerivedProduct: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'productCategory',
-    'productCode',
-    'status',
-    'request',
-    'quantity',
-    'parent',
-    'collection',
-    'processing',
-    'manipulation',
-    'storage'
-  ],
-  BodyStructure: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'morphology',
-    'location',
-    'locationQualifier',
-    'description',
-    'image'
-  ],
-  Bundle: ['id', 'meta', 'implicitRules', 'language', 'identifier', 'timestamp', 'total', 'link', 'entry', 'signature'],
-  CapabilityStatement: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'version',
-    'name',
-    'title',
-    'experimental',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'instantiates',
-    'imports',
-    'software',
-    'implementation',
-    'patchFormat',
-    'implementationGuide',
-    'rest',
-    'messaging',
-    'document'
-  ],
-  CarePlan: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'replaces',
-    'partOf',
-    'category',
-    'title',
-    'description',
-    'encounter',
-    'period',
-    'created',
-    'author',
-    'contributor',
-    'careTeam',
-    'addresses',
-    'supportingInfo',
-    'goal',
-    'activity',
-    'note'
-  ],
-  CareTeam: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'category',
-    'name',
-    'subject',
-    'encounter',
-    'period',
-    'participant',
-    'reasonCode',
-    'reasonReference',
-    'managingOrganization',
-    'telecom',
-    'note'
-  ],
-  CatalogEntry: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'additionalIdentifier',
-    'classification',
-    'status',
-    'validityPeriod',
-    'validTo',
-    'lastUpdated',
-    'additionalCharacteristic',
-    'additionalClassification',
-    'relatedEntry'
-  ],
-  ChargeItem: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'definitionUri',
-    'definitionCanonical',
-    'partOf',
-    'context',
-    'occurrence[x]',
-    'performer',
-    'performingOrganization',
-    'requestingOrganization',
-    'costCenter',
-    'quantity',
-    'bodysite',
-    'factorOverride',
-    'priceOverride',
-    'overrideReason',
-    'enterer',
-    'enteredDate',
-    'reason',
-    'service',
-    'product[x]',
-    'account',
-    'note',
-    'supportingInformation'
-  ],
-  ChargeItemDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'version',
-    'title',
-    'derivedFromUri',
-    'partOf',
-    'replaces',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'code',
-    'instance',
-    'applicability',
-    'propertyGroup'
-  ],
-  Claim: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subType',
-    'billablePeriod',
-    'enterer',
-    'insurer',
-    'fundsReserve',
-    'related',
-    'prescription',
-    'originalPrescription',
-    'payee',
-    'referral',
-    'facility',
-    'careTeam',
-    'supportingInfo',
-    'diagnosis',
-    'procedure',
-    'accident',
-    'item',
-    'total'
-  ],
-  ClaimResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subType',
-    'requestor',
-    'request',
-    'disposition',
-    'preAuthRef',
-    'preAuthPeriod',
-    'payeeType',
-    'item',
-    'addItem',
-    'adjudication',
-    'total',
-    'payment',
-    'fundsReserve',
-    'formCode',
-    'form',
-    'processNote',
-    'communicationRequest',
-    'insurance',
-    'error'
-  ],
-  ClinicalImpression: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'statusReason',
-    'code',
-    'description',
-    'encounter',
-    'effective[x]',
-    'date',
-    'assessor',
-    'previous',
-    'problem',
-    'investigation',
-    'protocol',
-    'summary',
-    'finding',
-    'prognosisCodeableConcept',
-    'prognosisReference',
-    'supportingInfo',
-    'note'
-  ],
-  CodeSystem: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'caseSensitive',
-    'valueSet',
-    'hierarchyMeaning',
-    'compositional',
-    'versionNeeded',
-    'supplements',
-    'count',
-    'filter',
-    'property',
-    'concept'
-  ],
-  Communication: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'partOf',
-    'inResponseTo',
-    'statusReason',
-    'category',
-    'priority',
-    'medium',
-    'subject',
-    'topic',
-    'about',
-    'encounter',
-    'sent',
-    'received',
-    'recipient',
-    'sender',
-    'reasonCode',
-    'reasonReference',
-    'payload',
-    'note'
-  ],
-  CommunicationRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'replaces',
-    'groupIdentifier',
-    'statusReason',
-    'category',
-    'priority',
-    'doNotPerform',
-    'medium',
-    'subject',
-    'about',
-    'encounter',
-    'payload',
-    'occurrence[x]',
-    'authoredOn',
-    'requester',
-    'recipient',
-    'sender',
-    'reasonCode',
-    'reasonReference',
-    'note'
-  ],
-  CompartmentDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'version',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'purpose',
-    'resource'
-  ],
-  Composition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'category',
-    'subject',
-    'encounter',
-    'confidentiality',
-    'attester',
-    'custodian',
-    'relatesTo',
-    'event',
-    'section'
-  ],
-  ConceptMap: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'source[x]',
-    'target[x]',
-    'group'
-  ],
-  Condition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'clinicalStatus',
-    'verificationStatus',
-    'category',
-    'severity',
-    'code',
-    'bodySite',
-    'encounter',
-    'onset[x]',
-    'abatement[x]',
-    'recordedDate',
-    'recorder',
-    'asserter',
-    'stage',
-    'evidence',
-    'note'
-  ],
-  Consent: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'patient',
-    'dateTime',
-    'performer',
-    'organization',
-    'source[x]',
-    'policy',
-    'policyRule',
-    'verification',
-    'provision'
-  ],
-  Contract: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'url',
-    'version',
-    'status',
-    'legalState',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'contentDerivative',
-    'issued',
-    'applies',
-    'expirationType',
-    'subject',
-    'authority',
-    'domain',
-    'site',
-    'name',
-    'title',
-    'subtitle',
-    'alias',
-    'author',
-    'scope',
-    'topic[x]',
-    'type',
-    'subType',
-    'contentDefinition',
-    'term',
-    'supportingInfo',
-    'relevantHistory',
-    'signer',
-    'friendly',
-    'legal',
-    'rule',
-    'legallyBinding[x]'
-  ],
-  Coverage: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'policyHolder',
-    'subscriber',
-    'subscriberId',
-    'dependent',
-    'relationship',
-    'period',
-    'class',
-    'order',
-    'network',
-    'costToBeneficiary',
-    'subrogation',
-    'contract'
-  ],
-  CoverageEligibilityRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'priority',
-    'serviced[x]',
-    'enterer',
-    'provider',
-    'facility',
-    'supportingInfo',
-    'insurance',
-    'item'
-  ],
-  CoverageEligibilityResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'serviced[x]',
-    'requestor',
-    'disposition',
-    'insurance',
-    'preAuthRef',
-    'form',
-    'error'
-  ],
-  DetectedIssue: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'code',
-    'severity',
-    'patient',
-    'identified[x]',
-    'author',
-    'implicated',
-    'evidence',
-    'detail',
-    'reference',
-    'mitigation'
-  ],
-  Device: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'definition',
-    'udiCarrier',
-    'status',
-    'statusReason',
-    'distinctIdentifier',
-    'manufacturer',
-    'manufactureDate',
-    'expirationDate',
-    'lotNumber',
-    'serialNumber',
-    'deviceName',
-    'modelNumber',
-    'partNumber',
-    'type',
-    'specialization',
-    'version',
-    'property',
-    'patient',
-    'owner',
-    'contact',
-    'location',
-    'url',
-    'note',
-    'safety',
-    'parent'
-  ],
-  DeviceDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'udiDeviceIdentifier',
-    'manufacturer[x]',
-    'deviceName',
-    'modelNumber',
-    'type',
-    'specialization',
-    'version',
-    'safety',
-    'shelfLifeStorage',
-    'physicalCharacteristics',
-    'languageCode',
-    'capability',
-    'property',
-    'owner',
-    'contact',
-    'url',
-    'onlineInformation',
-    'note',
-    'quantity',
-    'parentDevice',
-    'material'
-  ],
-  DeviceMetric: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'unit',
-    'source',
-    'parent',
-    'operationalStatus',
-    'color',
-    'measurementPeriod',
-    'calibration'
-  ],
-  DeviceRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'priorRequest',
-    'groupIdentifier',
-    'status',
-    'priority',
-    'parameter',
-    'encounter',
-    'occurrence[x]',
-    'authoredOn',
-    'requester',
-    'performerType',
-    'performer',
-    'reasonCode',
-    'reasonReference',
-    'insurance',
-    'supportingInfo',
-    'note',
-    'relevantHistory'
-  ],
-  DeviceUseStatement: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'derivedFrom',
-    'timing[x]',
-    'recordedOn',
-    'source',
-    'reasonCode',
-    'reasonReference',
-    'bodySite',
-    'note'
-  ],
-  DiagnosticReport: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'category',
-    'subject',
-    'encounter',
-    'effective[x]',
-    'issued',
-    'performer',
-    'resultsInterpreter',
-    'specimen',
-    'result',
-    'imagingStudy',
-    'media',
-    'conclusion',
-    'conclusionCode',
-    'presentedForm'
-  ],
-  DocumentManifest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'masterIdentifier',
-    'identifier',
-    'type',
-    'subject',
-    'created',
-    'author',
-    'recipient',
-    'source',
-    'description',
-    'related'
-  ],
-  DocumentReference: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'masterIdentifier',
-    'identifier',
-    'docStatus',
-    'type',
-    'category',
-    'subject',
-    'date',
-    'author',
-    'authenticator',
-    'custodian',
-    'relatesTo',
-    'description',
-    'securityLabel',
-    'context'
-  ],
-  EffectEvidenceSynthesis: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'note',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'synthesisType',
-    'studyType',
-    'sampleSize',
-    'resultsByExposure',
-    'effectEstimate',
-    'certainty'
-  ],
-  Encounter: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'statusHistory',
-    'classHistory',
-    'type',
-    'serviceType',
-    'priority',
-    'subject',
-    'episodeOfCare',
-    'basedOn',
-    'participant',
-    'appointment',
-    'period',
-    'length',
-    'reasonCode',
-    'reasonReference',
-    'diagnosis',
-    'account',
-    'hospitalization',
-    'location',
-    'serviceProvider',
-    'partOf'
-  ],
-  Endpoint: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'name',
-    'managingOrganization',
-    'contact',
-    'period',
-    'payloadMimeType',
-    'header'
-  ],
-  EnrollmentRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'created',
-    'insurer',
-    'provider',
-    'candidate',
-    'coverage'
-  ],
-  EnrollmentResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'request',
-    'outcome',
-    'disposition',
-    'created',
-    'organization',
-    'requestProvider'
-  ],
-  EpisodeOfCare: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'statusHistory',
-    'type',
-    'diagnosis',
-    'managingOrganization',
-    'period',
-    'referralRequest',
-    'careManager',
-    'team',
-    'account'
-  ],
-  EventDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact'
-  ],
-  Evidence: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'shortTitle',
-    'subtitle',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'note',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'exposureVariant',
-    'outcome'
-  ],
-  EvidenceVariable: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'shortTitle',
-    'subtitle',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'note',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'type'
-  ],
-  ExampleScenario: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'purpose',
-    'actor',
-    'instance',
-    'process',
-    'workflow'
-  ],
-  ExplanationOfBenefit: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subType',
-    'billablePeriod',
-    'enterer',
-    'priority',
-    'fundsReserveRequested',
-    'fundsReserve',
-    'related',
-    'prescription',
-    'originalPrescription',
-    'payee',
-    'referral',
-    'facility',
-    'claim',
-    'claimResponse',
-    'disposition',
-    'preAuthRef',
-    'preAuthRefPeriod',
-    'careTeam',
-    'supportingInfo',
-    'diagnosis',
-    'procedure',
-    'precedence',
-    'accident',
-    'item',
-    'addItem',
-    'adjudication',
-    'total',
-    'payment',
-    'formCode',
-    'form',
-    'processNote',
-    'benefitPeriod',
-    'benefitBalance'
-  ],
-  FamilyMemberHistory: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'dataAbsentReason',
-    'date',
-    'name',
-    'sex',
-    'born[x]',
-    'age[x]',
-    'estimatedAge',
-    'deceased[x]',
-    'reasonCode',
-    'reasonReference',
-    'note',
-    'condition'
-  ],
-  Flag: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'category',
-    'period',
-    'encounter',
-    'author'
-  ],
-  Goal: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'achievementStatus',
-    'category',
-    'priority',
-    'start[x]',
-    'target',
-    'statusDate',
-    'statusReason',
-    'expressedBy',
-    'addresses',
-    'note',
-    'outcomeCode',
-    'outcomeReference'
-  ],
-  GraphDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'version',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'profile',
-    'link'
-  ],
-  Group: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'code',
-    'name',
-    'quantity',
-    'managingEntity',
-    'characteristic',
-    'member'
-  ],
-  GuidanceResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'requestIdentifier',
-    'identifier',
-    'subject',
-    'encounter',
-    'occurrenceDateTime',
-    'performer',
-    'reasonCode',
-    'reasonReference',
-    'note',
-    'evaluationMessage',
-    'outputParameters',
-    'result',
-    'dataRequirement'
-  ],
-  HealthcareService: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'providedBy',
-    'category',
-    'type',
-    'specialty',
-    'location',
-    'name',
-    'comment',
-    'extraDetails',
-    'photo',
-    'telecom',
-    'coverageArea',
-    'serviceProvisionCode',
-    'eligibility',
-    'program',
-    'characteristic',
-    'communication',
-    'referralMethod',
-    'appointmentRequired',
-    'availableTime',
-    'notAvailable',
-    'availabilityExceptions',
-    'endpoint'
-  ],
-  ImagingStudy: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'modality',
-    'encounter',
-    'started',
-    'basedOn',
-    'referrer',
-    'interpreter',
-    'endpoint',
-    'numberOfSeries',
-    'numberOfInstances',
-    'procedureReference',
-    'procedureCode',
-    'location',
-    'reasonCode',
-    'reasonReference',
-    'note',
-    'description',
-    'series'
-  ],
-  Immunization: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'statusReason',
-    'encounter',
-    'recorded',
-    'primarySource',
-    'reportOrigin',
-    'location',
-    'manufacturer',
-    'lotNumber',
-    'expirationDate',
-    'site',
-    'route',
-    'doseQuantity',
-    'performer',
-    'note',
-    'reasonCode',
-    'reasonReference',
-    'isSubpotent',
-    'subpotentReason',
-    'education',
-    'programEligibility',
-    'fundingSource',
-    'reaction',
-    'protocolApplied'
-  ],
-  ImmunizationEvaluation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'date',
-    'authority',
-    'doseStatusReason',
-    'description',
-    'series',
-    'doseNumber[x]',
-    'seriesDoses[x]'
-  ],
-  ImmunizationRecommendation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'authority'
-  ],
-  ImplementationGuide: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'version',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'license',
-    'dependsOn',
-    'global',
-    'definition',
-    'manifest'
-  ],
-  InsurancePlan: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'type',
-    'name',
-    'alias',
-    'period',
-    'ownedBy',
-    'administeredBy',
-    'coverageArea',
-    'contact',
-    'endpoint',
-    'network',
-    'coverage',
-    'plan'
-  ],
-  Invoice: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'cancelledReason',
-    'type',
-    'subject',
-    'recipient',
-    'date',
-    'participant',
-    'issuer',
-    'account',
-    'lineItem',
-    'totalPriceComponent',
-    'totalNet',
-    'totalGross',
-    'paymentTerms',
-    'note'
-  ],
-  Library: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'parameter',
-    'dataRequirement',
-    'content'
-  ],
-  Linkage: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'active',
-    'author'
-  ],
-  List: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'title',
-    'code',
-    'subject',
-    'encounter',
-    'date',
-    'source',
-    'orderedBy',
-    'note',
-    'entry',
-    'emptyReason'
-  ],
-  Location: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'operationalStatus',
-    'name',
-    'alias',
-    'description',
-    'mode',
-    'type',
-    'telecom',
-    'address',
-    'physicalType',
-    'position',
-    'managingOrganization',
-    'partOf',
-    'hoursOfOperation',
-    'availabilityExceptions',
-    'endpoint'
-  ],
-  Measure: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'library',
-    'disclaimer',
-    'scoring',
-    'compositeScoring',
-    'type',
-    'riskAdjustment',
-    'rateAggregation',
-    'rationale',
-    'clinicalRecommendationStatement',
-    'improvementNotation',
-    'definition',
-    'guidance',
-    'group',
-    'supplementalData'
-  ],
-  MeasureReport: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subject',
-    'date',
-    'reporter',
-    'improvementNotation',
-    'group',
-    'evaluatedResource'
-  ],
-  Media: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'partOf',
-    'type',
-    'modality',
-    'view',
-    'subject',
-    'encounter',
-    'created[x]',
-    'issued',
-    'operator',
-    'reasonCode',
-    'bodySite',
-    'deviceName',
-    'device',
-    'height',
-    'width',
-    'frames',
-    'duration',
-    'note'
-  ],
-  Medication: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'code',
-    'status',
-    'manufacturer',
-    'form',
-    'amount',
-    'ingredient',
-    'batch'
-  ],
-  MedicationAdministration: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiates',
-    'partOf',
-    'statusReason',
-    'category',
-    'context',
-    'supportingInformation',
-    'performer',
-    'reasonCode',
-    'reasonReference',
-    'request',
-    'device',
-    'note',
-    'dosage',
-    'eventHistory'
-  ],
-  MedicationDispense: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'partOf',
-    'statusReason[x]',
-    'category',
-    'subject',
-    'context',
-    'supportingInformation',
-    'performer',
-    'location',
-    'authorizingPrescription',
-    'type',
-    'quantity',
-    'daysSupply',
-    'whenPrepared',
-    'whenHandedOver',
-    'destination',
-    'receiver',
-    'note',
-    'dosageInstruction',
-    'substitution',
-    'detectedIssue',
-    'eventHistory'
-  ],
-  MedicationKnowledge: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'code',
-    'status',
-    'manufacturer',
-    'doseForm',
-    'amount',
-    'synonym',
-    'relatedMedicationKnowledge',
-    'associatedMedication',
-    'productType',
-    'monograph',
-    'ingredient',
-    'preparationInstruction',
-    'intendedRoute',
-    'cost',
-    'monitoringProgram',
-    'administrationGuidelines',
-    'medicineClassification',
-    'packaging',
-    'drugCharacteristic',
-    'contraindication',
-    'regulatory',
-    'kinetics'
-  ],
-  MedicationRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'statusReason',
-    'category',
-    'priority',
-    'doNotPerform',
-    'reported[x]',
-    'encounter',
-    'supportingInformation',
-    'authoredOn',
-    'requester',
-    'performer',
-    'performerType',
-    'recorder',
-    'reasonCode',
-    'reasonReference',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'groupIdentifier',
-    'courseOfTherapyType',
-    'insurance',
-    'note',
-    'dosageInstruction',
-    'dispenseRequest',
-    'substitution',
-    'priorPrescription',
-    'detectedIssue',
-    'eventHistory'
-  ],
-  MedicationStatement: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'partOf',
-    'statusReason',
-    'category',
-    'context',
-    'effective[x]',
-    'dateAsserted',
-    'informationSource',
-    'derivedFrom',
-    'reasonCode',
-    'reasonReference',
-    'note',
-    'dosage'
-  ],
-  MedicinalProduct: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'domain',
-    'combinedPharmaceuticalDoseForm',
-    'legalStatusOfSupply',
-    'additionalMonitoringIndicator',
-    'specialMeasures',
-    'paediatricUseIndicator',
-    'productClassification',
-    'marketingStatus',
-    'pharmaceuticalProduct',
-    'packagedMedicinalProduct',
-    'attachedDocument',
-    'masterFile',
-    'contact',
-    'clinicalTrial',
-    'crossReference',
-    'manufacturingBusinessOperation',
-    'specialDesignation'
-  ],
-  MedicinalProductAuthorization: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subject',
-    'country',
-    'jurisdiction',
-    'status',
-    'statusDate',
-    'restoreDate',
-    'validityPeriod',
-    'dataExclusivityPeriod',
-    'dateOfFirstAuthorization',
-    'internationalBirthDate',
-    'legalBasis',
-    'jurisdictionalAuthorization',
-    'holder',
-    'regulator',
-    'procedure'
-  ],
-  MedicinalProductContraindication: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'subject',
-    'disease',
-    'diseaseStatus',
-    'comorbidity',
-    'therapeuticIndication',
-    'otherTherapy',
-    'population'
-  ],
-  MedicinalProductIndication: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'subject',
-    'diseaseSymptomProcedure',
-    'diseaseStatus',
-    'comorbidity',
-    'intendedEffect',
-    'duration',
-    'otherTherapy',
-    'undesirableEffect',
-    'population'
-  ],
-  MedicinalProductIngredient: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'allergenicIndicator',
-    'manufacturer',
-    'specifiedSubstance',
-    'substance'
-  ],
-  MedicinalProductInteraction: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'subject',
-    'description',
-    'interactant',
-    'type',
-    'effect',
-    'incidence',
-    'management'
-  ],
-  MedicinalProductManufactured: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'unitOfPresentation',
-    'manufacturer',
-    'ingredient',
-    'physicalCharacteristics',
-    'otherCharacteristics'
-  ],
-  MedicinalProductPackaged: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'subject',
-    'description',
-    'legalStatusOfSupply',
-    'marketingStatus',
-    'marketingAuthorization',
-    'manufacturer',
-    'batchIdentifier'
-  ],
-  MedicinalProductPharmaceutical: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'unitOfPresentation',
-    'ingredient',
-    'device',
-    'characteristics'
-  ],
-  MedicinalProductUndesirableEffect: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'subject',
-    'symptomConditionEffect',
-    'classification',
-    'frequencyOfOccurrence',
-    'population'
-  ],
-  MessageDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'replaces',
-    'experimental',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'base',
-    'parent',
-    'category',
-    'focus',
-    'responseRequired',
-    'allowedResponse',
-    'graph'
-  ],
-  MessageHeader: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'destination',
-    'sender',
-    'enterer',
-    'author',
-    'responsible',
-    'reason',
-    'response',
-    'focus',
-    'definition'
-  ],
-  MolecularSequence: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'patient',
-    'specimen',
-    'device',
-    'performer',
-    'quantity',
-    'referenceSeq',
-    'variant',
-    'observedSeq',
-    'quality',
-    'readCoverage',
-    'repository',
-    'pointer',
-    'structureVariant'
-  ],
-  NamingSystem: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'publisher',
-    'contact',
-    'responsible',
-    'type',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'usage'
-  ],
-  NutritionOrder: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'instantiates',
-    'encounter',
-    'orderer',
-    'allergyIntolerance',
-    'foodPreferenceModifier',
-    'excludeFoodModifier',
-    'oralDiet',
-    'supplement',
-    'enteralFormula',
-    'note'
-  ],
-  Observation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'partOf',
-    'category',
-    'subject',
-    'focus',
-    'encounter',
-    'effective[x]',
-    'issued',
-    'performer',
-    'value[x]',
-    'dataAbsentReason',
-    'interpretation',
-    'note',
-    'bodySite',
-    'method',
-    'specimen',
-    'device',
-    'referenceRange',
-    'hasMember',
-    'derivedFrom',
-    'component'
-  ],
-  ObservationDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'category',
-    'identifier',
-    'permittedDataType',
-    'multipleResultsAllowed',
-    'method',
-    'preferredReportName',
-    'quantitativeDetails',
-    'qualifiedInterval',
-    'validCodedValueSet',
-    'normalCodedValueSet',
-    'abnormalCodedValueSet',
-    'criticalCodedValueSet'
-  ],
-  OperationDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'version',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'affectsState',
-    'comment',
-    'base',
-    'resource',
-    'inputProfile',
-    'outputProfile',
-    'parameter',
-    'overload'
-  ],
-  OperationOutcome: ['id', 'meta', 'implicitRules', 'language', 'text', 'contained', 'extension', 'modifierExtension'],
-  Organization: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'type',
-    'name',
-    'alias',
-    'telecom',
-    'address',
-    'partOf',
-    'contact',
-    'endpoint'
-  ],
-  OrganizationAffiliation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'period',
-    'organization',
-    'participatingOrganization',
-    'network',
-    'code',
-    'specialty',
-    'location',
-    'healthcareService',
-    'telecom',
-    'endpoint'
-  ],
-  Parameters: ['id', 'meta', 'implicitRules', 'language', 'parameter'],
-  Patient: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'name',
-    'telecom',
-    'gender',
-    'birthDate',
-    'deceased[x]',
-    'address',
-    'maritalStatus',
-    'multipleBirth[x]',
-    'photo',
-    'contact',
-    'communication',
-    'generalPractitioner',
-    'managingOrganization',
-    'link'
-  ],
-  PaymentNotice: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'request',
-    'response',
-    'provider',
-    'paymentDate',
-    'payee',
-    'paymentStatus'
-  ],
-  PaymentReconciliation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'period',
-    'paymentIssuer',
-    'request',
-    'requestor',
-    'outcome',
-    'disposition',
-    'paymentIdentifier',
-    'detail',
-    'formCode',
-    'processNote'
-  ],
-  Person: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'name',
-    'telecom',
-    'gender',
-    'birthDate',
-    'address',
-    'photo',
-    'managingOrganization',
-    'active',
-    'link'
-  ],
-  PlanDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'subtitle',
-    'type',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'library',
-    'goal',
-    'action'
-  ],
-  Practitioner: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'name',
-    'telecom',
-    'address',
-    'gender',
-    'birthDate',
-    'photo',
-    'qualification',
-    'communication'
-  ],
-  PractitionerRole: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'period',
-    'practitioner',
-    'organization',
-    'code',
-    'specialty',
-    'location',
-    'healthcareService',
-    'telecom',
-    'availableTime',
-    'notAvailable',
-    'availabilityExceptions',
-    'endpoint'
-  ],
-  Procedure: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'partOf',
-    'statusReason',
-    'category',
-    'code',
-    'encounter',
-    'performed[x]',
-    'recorder',
-    'asserter',
-    'performer',
-    'location',
-    'reasonCode',
-    'reasonReference',
-    'bodySite',
-    'outcome',
-    'report',
-    'complication',
-    'complicationDetail',
-    'followUp',
-    'note',
-    'focalDevice',
-    'usedReference',
-    'usedCode'
-  ],
-  Provenance: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'occurred[x]',
-    'policy',
-    'location',
-    'reason',
-    'activity',
-    'entity',
-    'signature'
-  ],
-  Questionnaire: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'derivedFrom',
-    'experimental',
-    'subjectType',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'code',
-    'item'
-  ],
-  QuestionnaireResponse: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'partOf',
-    'questionnaire',
-    'subject',
-    'encounter',
-    'authored',
-    'author',
-    'source',
-    'item'
-  ],
-  RelatedPerson: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'relationship',
-    'name',
-    'telecom',
-    'gender',
-    'birthDate',
-    'address',
-    'photo',
-    'period',
-    'communication'
-  ],
-  RequestGroup: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'replaces',
-    'groupIdentifier',
-    'priority',
-    'code',
-    'subject',
-    'encounter',
-    'authoredOn',
-    'author',
-    'reasonCode',
-    'reasonReference',
-    'note',
-    'action'
-  ],
-  ResearchDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'shortTitle',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'comment',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'library',
-    'exposure',
-    'exposureAlternative',
-    'outcome'
-  ],
-  ResearchElementDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'shortTitle',
-    'subtitle',
-    'experimental',
-    'subject[x]',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'comment',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'usage',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'library',
-    'variableType'
-  ],
-  ResearchStudy: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'title',
-    'protocol',
-    'partOf',
-    'primaryPurposeType',
-    'phase',
-    'category',
-    'focus',
-    'condition',
-    'contact',
-    'relatedArtifact',
-    'keyword',
-    'location',
-    'description',
-    'enrollment',
-    'period',
-    'sponsor',
-    'principalInvestigator',
-    'site',
-    'reasonStopped',
-    'note',
-    'arm',
-    'objective'
-  ],
-  ResearchSubject: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'period',
-    'assignedArm',
-    'actualArm',
-    'consent'
-  ],
-  RiskAssessment: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'parent',
-    'method',
-    'code',
-    'encounter',
-    'occurrence[x]',
-    'condition',
-    'performer',
-    'reasonCode',
-    'reasonReference',
-    'basis',
-    'prediction',
-    'mitigation',
-    'note'
-  ],
-  RiskEvidenceSynthesis: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'note',
-    'useContext',
-    'jurisdiction',
-    'copyright',
-    'approvalDate',
-    'lastReviewDate',
-    'effectivePeriod',
-    'topic',
-    'author',
-    'editor',
-    'reviewer',
-    'endorser',
-    'relatedArtifact',
-    'synthesisType',
-    'studyType',
-    'exposure',
-    'sampleSize',
-    'riskEstimate',
-    'certainty'
-  ],
-  Schedule: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'active',
-    'serviceCategory',
-    'serviceType',
-    'specialty',
-    'planningHorizon',
-    'comment'
-  ],
-  SearchParameter: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'version',
-    'derivedFrom',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'expression',
-    'xpath',
-    'xpathUsage',
-    'target',
-    'multipleOr',
-    'multipleAnd',
-    'comparator',
-    'modifier',
-    'chain',
-    'component'
-  ],
-  ServiceRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'replaces',
-    'requisition',
-    'category',
-    'priority',
-    'doNotPerform',
-    'code',
-    'orderDetail',
-    'quantity[x]',
-    'encounter',
-    'occurrence[x]',
-    'asNeeded[x]',
-    'authoredOn',
-    'requester',
-    'performerType',
-    'performer',
-    'locationCode',
-    'locationReference',
-    'reasonCode',
-    'reasonReference',
-    'insurance',
-    'supportingInfo',
-    'specimen',
-    'bodySite',
-    'note',
-    'patientInstruction',
-    'relevantHistory'
-  ],
-  Slot: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'serviceCategory',
-    'serviceType',
-    'specialty',
-    'appointmentType',
-    'overbooked',
-    'comment'
-  ],
-  Specimen: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'accessionIdentifier',
-    'status',
-    'type',
-    'subject',
-    'receivedTime',
-    'parent',
-    'request',
-    'collection',
-    'processing',
-    'container',
-    'condition',
-    'note'
-  ],
-  SpecimenDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'typeCollected',
-    'patientPreparation',
-    'timeAspect',
-    'collection',
-    'typeTested'
-  ],
-  StructureDefinition: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'version',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'keyword',
-    'fhirVersion',
-    'mapping',
-    'context',
-    'contextInvariant',
-    'baseDefinition',
-    'derivation',
-    'snapshot',
-    'differential'
-  ],
-  StructureMap: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'version',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'structure',
-    'import'
-  ],
-  Subscription: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'contact',
-    'end',
-    'error'
-  ],
-  Substance: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'category',
-    'description',
-    'instance',
-    'ingredient'
-  ],
-  SubstancePolymer: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'class',
-    'geometry',
-    'copolymerConnectivity',
-    'modification',
-    'monomerSet',
-    'repeat'
-  ],
-  SubstanceProtein: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'sequenceType',
-    'numberOfSubunits',
-    'disulfideLinkage',
-    'subunit'
-  ],
-  SubstanceReferenceInformation: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'comment',
-    'gene',
-    'geneElement',
-    'classification',
-    'target'
-  ],
-  SubstanceSourceMaterial: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'sourceMaterialClass',
-    'sourceMaterialType',
-    'sourceMaterialState',
-    'organismId',
-    'organismName',
-    'parentSubstanceId',
-    'parentSubstanceName',
-    'countryOfOrigin',
-    'geographicalLocation',
-    'developmentStage',
-    'fractionDescription',
-    'organism',
-    'partDescription'
-  ],
-  SubstanceSpecification: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'type',
-    'status',
-    'domain',
-    'description',
-    'source',
-    'comment',
-    'moiety',
-    'property',
-    'referenceInformation',
-    'structure',
-    'code',
-    'name',
-    'molecularWeight',
-    'relationship',
-    'nucleicAcid',
-    'polymer',
-    'protein',
-    'sourceMaterial'
-  ],
-  SupplyDelivery: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'basedOn',
-    'partOf',
-    'status',
-    'patient',
-    'type',
-    'suppliedItem',
-    'occurrence[x]',
-    'supplier',
-    'destination',
-    'receiver'
-  ],
-  SupplyRequest: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'status',
-    'category',
-    'priority',
-    'parameter',
-    'occurrence[x]',
-    'authoredOn',
-    'requester',
-    'supplier',
-    'reasonCode',
-    'reasonReference',
-    'deliverFrom',
-    'deliverTo'
-  ],
-  Task: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'instantiatesCanonical',
-    'instantiatesUri',
-    'basedOn',
-    'groupIdentifier',
-    'partOf',
-    'statusReason',
-    'businessStatus',
-    'priority',
-    'code',
-    'description',
-    'focus',
-    'for',
-    'encounter',
-    'executionPeriod',
-    'authoredOn',
-    'lastModified',
-    'requester',
-    'performerType',
-    'owner',
-    'location',
-    'reasonCode',
-    'reasonReference',
-    'insurance',
-    'note',
-    'relevantHistory',
-    'restriction',
-    'input',
-    'output'
-  ],
-  TerminologyCapabilities: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'version',
-    'name',
-    'title',
-    'experimental',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'software',
-    'implementation',
-    'lockedDate',
-    'codeSystem',
-    'expansion',
-    'codeSearch',
-    'validateCode',
-    'translation',
-    'closure'
-  ],
-  TestReport: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'name',
-    'score',
-    'tester',
-    'issued',
-    'participant',
-    'setup',
-    'test',
-    'teardown'
-  ],
-  TestScript: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'version',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'purpose',
-    'copyright',
-    'origin',
-    'destination',
-    'metadata',
-    'fixture',
-    'profile',
-    'variable',
-    'setup',
-    'test',
-    'teardown'
-  ],
-  ValueSet: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'url',
-    'identifier',
-    'version',
-    'name',
-    'title',
-    'experimental',
-    'date',
-    'publisher',
-    'contact',
-    'description',
-    'useContext',
-    'jurisdiction',
-    'immutable',
-    'purpose',
-    'copyright',
-    'compose',
-    'expansion'
-  ],
-  VerificationResult: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'target',
-    'targetLocation',
-    'need',
-    'statusDate',
-    'validationType',
-    'validationProcess',
-    'frequency',
-    'lastPerformed',
-    'nextScheduled',
-    'failureAction',
-    'primarySource',
-    'attestation',
-    'validator'
-  ],
-  VisionPrescription: [
-    'id',
-    'meta',
-    'implicitRules',
-    'language',
-    'text',
-    'contained',
-    'extension',
-    'modifierExtension',
-    'identifier',
-    'encounter'
-  ]
-};
