+++ src/data/propertyPaths.ts
@@ -1,5 +1,13 @@
 export const parsedPropertyPaths: Record<string, string[]> = {
   Account: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -13,6 +21,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'partOf'
   ],
   ActivityDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -21,7 +37,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subtitle',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -47,10 +63,10 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'intent',
     'priority',
     'doNotPerform',
-    'timing',
+    'timing[x]',
     'location',
     'participant',
-    'product',
+    'product[x]',
     'quantity',
     'dosage',
     'bodySite',
@@ -61,6 +77,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'dynamicValue'
   ],
   AdverseEvent: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'actuality',
     'category',
@@ -83,6 +107,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'study'
   ],
   AllergyIntolerance: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'clinicalStatus',
     'verificationStatus',
@@ -92,7 +124,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'code',
     'patient',
     'encounter',
-    'onset',
+    'onset[x]',
     'recordedDate',
     'recorder',
     'asserter',
@@ -101,6 +133,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'reaction'
   ],
   Appointment: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'cancelationReason',
@@ -125,6 +165,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'requestedPeriod'
   ],
   AppointmentResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'appointment',
     'start',
@@ -135,6 +183,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'comment'
   ],
   AuditEvent: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'type',
     'subtype',
     'action',
@@ -147,8 +203,31 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'source',
     'entity'
   ],
-  Basic: ['identifier', 'code', 'subject', 'created', 'author'],
+  Basic: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'code',
+    'subject',
+    'created',
+    'author'
+  ],
+  Binary: ['id', 'meta', 'implicitRules', 'language', 'contentType', 'securityContext', 'data'],
   BiologicallyDerivedProduct: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'productCategory',
     'productCode',
@@ -162,6 +241,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'storage'
   ],
   BodyStructure: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'morphology',
@@ -171,7 +258,28 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'image',
     'patient'
   ],
+  Bundle: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'identifier',
+    'type',
+    'timestamp',
+    'total',
+    'link',
+    'entry',
+    'signature'
+  ],
   CapabilityStatement: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -200,6 +308,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'document'
   ],
   CarePlan: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -225,6 +341,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   CareTeam: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'category',
@@ -240,6 +364,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   CatalogEntry: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'type',
     'orderable',
@@ -255,6 +387,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'relatedEntry'
   ],
   ChargeItem: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'definitionUri',
     'definitionCanonical',
@@ -263,7 +403,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'code',
     'subject',
     'context',
-    'occurrence',
+    'occurrence[x]',
     'performer',
     'performingOrganization',
     'requestingOrganization',
@@ -277,12 +417,20 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'enteredDate',
     'reason',
     'service',
-    'product',
+    'product[x]',
     'account',
     'note',
     'supportingInformation'
   ],
   ChargeItemDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -308,6 +456,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'propertyGroup'
   ],
   Claim: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -337,6 +493,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'total'
   ],
   ClaimResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -366,6 +530,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'error'
   ],
   ClinicalImpression: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'statusReason',
@@ -373,7 +545,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'description',
     'subject',
     'encounter',
-    'effective',
+    'effective[x]',
     'date',
     'assessor',
     'previous',
@@ -388,6 +560,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   CodeSystem: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -416,6 +596,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'concept'
   ],
   Communication: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -441,6 +629,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   CommunicationRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'replaces',
@@ -455,7 +651,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'about',
     'encounter',
     'payload',
-    'occurrence',
+    'occurrence[x]',
     'authoredOn',
     'requester',
     'recipient',
@@ -465,6 +661,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   CompartmentDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -481,6 +685,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'resource'
   ],
   Composition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -498,6 +710,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'section'
   ],
   ConceptMap: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -513,11 +733,19 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'jurisdiction',
     'purpose',
     'copyright',
-    'source',
-    'target',
+    'source[x]',
+    'target[x]',
     'group'
   ],
   Condition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'clinicalStatus',
     'verificationStatus',
@@ -527,8 +755,8 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'bodySite',
     'subject',
     'encounter',
-    'onset',
-    'abatement',
+    'onset[x]',
+    'abatement[x]',
     'recordedDate',
     'recorder',
     'asserter',
@@ -537,6 +765,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   Consent: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'scope',
@@ -545,13 +781,21 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'dateTime',
     'performer',
     'organization',
-    'source',
+    'source[x]',
     'policy',
     'policyRule',
     'verification',
     'provision'
   ],
   Contract: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'url',
     'version',
@@ -573,7 +817,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'alias',
     'author',
     'scope',
-    'topic',
+    'topic[x]',
     'type',
     'subType',
     'contentDefinition',
@@ -584,9 +828,17 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'friendly',
     'legal',
     'rule',
-    'legallyBinding'
+    'legallyBinding[x]'
   ],
   Coverage: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -606,12 +858,20 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'contract'
   ],
   CoverageEligibilityRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'priority',
     'purpose',
     'patient',
-    'serviced',
+    'serviced[x]',
     'created',
     'enterer',
     'provider',
@@ -622,11 +882,19 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'item'
   ],
   CoverageEligibilityResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'purpose',
     'patient',
-    'serviced',
+    'serviced[x]',
     'created',
     'requestor',
     'request',
@@ -639,12 +907,20 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'error'
   ],
   DetectedIssue: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'code',
     'severity',
     'patient',
-    'identified',
+    'identified[x]',
     'author',
     'implicated',
     'evidence',
@@ -653,6 +929,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'mitigation'
   ],
   Device: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'definition',
     'udiCarrier',
@@ -681,9 +965,17 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'parent'
   ],
   DeviceDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'udiDeviceIdentifier',
-    'manufacturer',
+    'manufacturer[x]',
     'deviceName',
     'modelNumber',
     'type',
@@ -705,6 +997,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'material'
   ],
   DeviceMetric: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'type',
     'unit',
@@ -717,6 +1017,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'calibration'
   ],
   DeviceRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -726,11 +1034,11 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'status',
     'intent',
     'priority',
-    'code',
+    'code[x]',
     'parameter',
     'subject',
     'encounter',
-    'occurrence',
+    'occurrence[x]',
     'authoredOn',
     'requester',
     'performerType',
@@ -743,12 +1051,20 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'relevantHistory'
   ],
   DeviceUseStatement: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'status',
     'subject',
     'derivedFrom',
-    'timing',
+    'timing[x]',
     'recordedOn',
     'source',
     'device',
@@ -758,6 +1074,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   DiagnosticReport: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'status',
@@ -765,7 +1089,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'code',
     'subject',
     'encounter',
-    'effective',
+    'effective[x]',
     'issued',
     'performer',
     'resultsInterpreter',
@@ -778,6 +1102,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'presentedForm'
   ],
   DocumentManifest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'masterIdentifier',
     'identifier',
     'status',
@@ -792,6 +1124,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'related'
   ],
   DocumentReference: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'masterIdentifier',
     'identifier',
     'status',
@@ -810,6 +1150,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'context'
   ],
   EffectEvidenceSynthesis: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -845,6 +1193,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'certainty'
   ],
   Encounter: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'statusHistory',
@@ -870,6 +1226,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'partOf'
   ],
   Endpoint: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'connectionType',
@@ -882,8 +1246,32 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'address',
     'header'
   ],
-  EnrollmentRequest: ['identifier', 'status', 'created', 'insurer', 'provider', 'candidate', 'coverage'],
+  EnrollmentRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'status',
+    'created',
+    'insurer',
+    'provider',
+    'candidate',
+    'coverage'
+  ],
   EnrollmentResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'request',
@@ -894,6 +1282,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'requestProvider'
   ],
   EpisodeOfCare: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'statusHistory',
@@ -908,6 +1304,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'account'
   ],
   EventDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -916,7 +1320,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subtitle',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -938,6 +1342,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'trigger'
   ],
   Evidence: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -968,6 +1380,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'outcome'
   ],
   EvidenceVariable: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -997,6 +1417,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'characteristic'
   ],
   ExampleScenario: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -1016,6 +1444,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'workflow'
   ],
   ExplanationOfBenefit: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -1061,6 +1497,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'benefitBalance'
   ],
   FamilyMemberHistory: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -1071,17 +1515,42 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'name',
     'relationship',
     'sex',
-    'born',
-    'age',
+    'born[x]',
+    'age[x]',
     'estimatedAge',
-    'deceased',
+    'deceased[x]',
     'reasonCode',
     'reasonReference',
     'note',
     'condition'
   ],
-  Flag: ['identifier', 'status', 'category', 'code', 'subject', 'period', 'encounter', 'author'],
+  Flag: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'status',
+    'category',
+    'code',
+    'subject',
+    'period',
+    'encounter',
+    'author'
+  ],
   Goal: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'lifecycleStatus',
     'achievementStatus',
@@ -1089,7 +1558,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'priority',
     'description',
     'subject',
-    'start',
+    'start[x]',
     'target',
     'statusDate',
     'statusReason',
@@ -1100,6 +1569,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'outcomeReference'
   ],
   GraphDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -1117,6 +1594,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'link'
   ],
   Group: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'type',
@@ -1129,9 +1614,17 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'member'
   ],
   GuidanceResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'requestIdentifier',
     'identifier',
-    'module',
+    'module[x]',
     'status',
     'subject',
     'encounter',
@@ -1146,6 +1639,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'dataRequirement'
   ],
   HealthcareService: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'providedBy',
@@ -1172,6 +1673,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'endpoint'
   ],
   ImagingStudy: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'modality',
@@ -1194,13 +1703,21 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'series'
   ],
   Immunization: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'statusReason',
     'vaccineCode',
     'patient',
     'encounter',
-    'occurrence',
+    'occurrence[x]',
     'recorded',
     'primarySource',
     'reportOrigin',
@@ -1224,6 +1741,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'protocolApplied'
   ],
   ImmunizationEvaluation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'patient',
@@ -1235,11 +1760,33 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'doseStatusReason',
     'description',
     'series',
-    'doseNumber',
-    'seriesDoses'
+    'doseNumber[x]',
+    'seriesDoses[x]'
   ],
-  ImmunizationRecommendation: ['identifier', 'patient', 'date', 'authority', 'recommendation'],
+  ImmunizationRecommendation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'patient',
+    'date',
+    'authority',
+    'recommendation'
+  ],
   ImplementationGuide: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -1262,6 +1809,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'manifest'
   ],
   InsurancePlan: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -1278,6 +1833,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'plan'
   ],
   Invoice: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'cancelledReason',
@@ -1296,6 +1859,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   Library: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -1305,7 +1876,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'status',
     'experimental',
     'type',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -1328,8 +1899,28 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'dataRequirement',
     'content'
   ],
-  Linkage: ['active', 'author', 'item'],
+  Linkage: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'active',
+    'author',
+    'item'
+  ],
   List: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'mode',
@@ -1345,6 +1936,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'emptyReason'
   ],
   Location: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'operationalStatus',
@@ -1364,6 +1963,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'endpoint'
   ],
   Measure: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -1372,7 +1979,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subtitle',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -1407,6 +2014,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'supplementalData'
   ],
   MeasureReport: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'type',
@@ -1420,6 +2035,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'evaluatedResource'
   ],
   Media: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'partOf',
@@ -1429,7 +2052,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'view',
     'subject',
     'encounter',
-    'created',
+    'created[x]',
     'issued',
     'operator',
     'reasonCode',
@@ -1443,19 +2066,44 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'content',
     'note'
   ],
-  Medication: ['identifier', 'code', 'status', 'manufacturer', 'form', 'amount', 'ingredient', 'batch'],
+  Medication: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'code',
+    'status',
+    'manufacturer',
+    'form',
+    'amount',
+    'ingredient',
+    'batch'
+  ],
   MedicationAdministration: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiates',
     'partOf',
     'status',
     'statusReason',
     'category',
-    'medication',
+    'medication[x]',
     'subject',
     'context',
     'supportingInformation',
-    'effective',
+    'effective[x]',
     'performer',
     'reasonCode',
     'reasonReference',
@@ -1466,12 +2114,20 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'eventHistory'
   ],
   MedicationDispense: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'partOf',
     'status',
-    'statusReason',
+    'statusReason[x]',
     'category',
-    'medication',
+    'medication[x]',
     'subject',
     'context',
     'supportingInformation',
@@ -1492,6 +2148,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'eventHistory'
   ],
   MedicationKnowledge: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'code',
     'status',
     'manufacturer',
@@ -1516,6 +2180,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'kinetics'
   ],
   MedicationRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'statusReason',
@@ -1523,8 +2195,8 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'category',
     'priority',
     'doNotPerform',
-    'reported',
-    'medication',
+    'reported[x]',
+    'medication[x]',
     'subject',
     'encounter',
     'supportingInformation',
@@ -1550,16 +2222,24 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'eventHistory'
   ],
   MedicationStatement: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'partOf',
     'status',
     'statusReason',
     'category',
-    'medication',
+    'medication[x]',
     'subject',
     'context',
-    'effective',
+    'effective[x]',
     'dateAsserted',
     'informationSource',
     'derivedFrom',
@@ -1569,6 +2249,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'dosage'
   ],
   MedicinalProduct: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'type',
     'domain',
@@ -1591,6 +2279,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'specialDesignation'
   ],
   MedicinalProductAuthorization: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'subject',
     'country',
@@ -1609,6 +2305,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'procedure'
   ],
   MedicinalProductContraindication: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'subject',
     'disease',
     'diseaseStatus',
@@ -1618,6 +2322,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'population'
   ],
   MedicinalProductIndication: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'subject',
     'diseaseSymptomProcedure',
     'diseaseStatus',
@@ -1629,6 +2341,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'population'
   ],
   MedicinalProductIngredient: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'role',
     'allergenicIndicator',
@@ -1636,8 +2356,32 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'specifiedSubstance',
     'substance'
   ],
-  MedicinalProductInteraction: ['subject', 'description', 'interactant', 'type', 'effect', 'incidence', 'management'],
+  MedicinalProductInteraction: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'subject',
+    'description',
+    'interactant',
+    'type',
+    'effect',
+    'incidence',
+    'management'
+  ],
   MedicinalProductManufactured: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'manufacturedDoseForm',
     'unitOfPresentation',
     'quantity',
@@ -1647,6 +2391,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'otherCharacteristics'
   ],
   MedicinalProductPackaged: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'subject',
     'description',
@@ -1658,6 +2410,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'packageItem'
   ],
   MedicinalProductPharmaceutical: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'administrableDoseForm',
     'unitOfPresentation',
@@ -1667,6 +2427,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'routeOfAdministration'
   ],
   MedicinalProductUndesirableEffect: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'subject',
     'symptomConditionEffect',
     'classification',
@@ -1674,6 +2442,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'population'
   ],
   MessageDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -1692,7 +2468,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'copyright',
     'base',
     'parent',
-    'event',
+    'event[x]',
     'category',
     'focus',
     'responseRequired',
@@ -1700,7 +2476,15 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'graph'
   ],
   MessageHeader: [
-    'event',
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'event[x]',
     'destination',
     'sender',
     'enterer',
@@ -1713,6 +2497,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'definition'
   ],
   MolecularSequence: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'type',
     'coordinateSystem',
@@ -1731,6 +2523,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'structureVariant'
   ],
   NamingSystem: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'name',
     'status',
     'kind',
@@ -1746,6 +2546,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'uniqueId'
   ],
   NutritionOrder: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -1765,6 +2573,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   Observation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'partOf',
@@ -1774,10 +2590,10 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subject',
     'focus',
     'encounter',
-    'effective',
+    'effective[x]',
     'issued',
     'performer',
-    'value',
+    'value[x]',
     'dataAbsentReason',
     'interpretation',
     'note',
@@ -1791,6 +2607,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'component'
   ],
   ObservationDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'category',
     'code',
     'identifier',
@@ -1806,6 +2630,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'criticalCodedValueSet'
   ],
   OperationDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -1833,8 +2665,26 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'parameter',
     'overload'
   ],
-  OperationOutcome: ['issue'],
+  OperationOutcome: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'issue'
+  ],
   Organization: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'type',
@@ -1847,6 +2697,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'endpoint'
   ],
   OrganizationAffiliation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'period',
@@ -1860,17 +2718,26 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'telecom',
     'endpoint'
   ],
+  Parameters: ['id', 'meta', 'implicitRules', 'language', 'parameter'],
   Patient: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'name',
     'telecom',
     'gender',
     'birthDate',
-    'deceased',
+    'deceased[x]',
     'address',
     'maritalStatus',
-    'multipleBirth',
+    'multipleBirth[x]',
     'photo',
     'contact',
     'communication',
@@ -1879,6 +2746,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'link'
   ],
   PaymentNotice: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'request',
@@ -1893,6 +2768,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'paymentStatus'
   ],
   PaymentReconciliation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'period',
@@ -1910,6 +2793,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'processNote'
   ],
   Person: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'name',
     'telecom',
@@ -1922,6 +2813,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'link'
   ],
   PlanDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -1931,7 +2830,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'type',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -1955,6 +2854,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'action'
   ],
   Practitioner: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'name',
@@ -1967,6 +2874,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'communication'
   ],
   PractitionerRole: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'period',
@@ -1983,6 +2898,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'endpoint'
   ],
   Procedure: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -1994,7 +2917,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'code',
     'subject',
     'encounter',
-    'performed',
+    'performed[x]',
     'recorder',
     'asserter',
     'performer',
@@ -2013,8 +2936,16 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'usedCode'
   ],
   Provenance: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'target',
-    'occurred',
+    'occurred[x]',
     'recorded',
     'policy',
     'location',
@@ -2025,6 +2956,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'signature'
   ],
   Questionnaire: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2049,6 +2988,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'item'
   ],
   QuestionnaireResponse: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'partOf',
@@ -2062,6 +3009,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'item'
   ],
   RelatedPerson: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'patient',
@@ -2076,6 +3031,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'communication'
   ],
   RequestGroup: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -2096,6 +3059,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'action'
   ],
   ResearchDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2105,7 +3076,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subtitle',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -2132,6 +3103,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'outcome'
   ],
   ResearchElementDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2141,7 +3120,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'subtitle',
     'status',
     'experimental',
-    'subject',
+    'subject[x]',
     'date',
     'publisher',
     'contact',
@@ -2167,6 +3146,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'characteristic'
   ],
   ResearchStudy: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'title',
     'protocol',
@@ -2192,8 +3179,33 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'arm',
     'objective'
   ],
-  ResearchSubject: ['identifier', 'status', 'period', 'study', 'individual', 'assignedArm', 'actualArm', 'consent'],
+  ResearchSubject: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'status',
+    'period',
+    'study',
+    'individual',
+    'assignedArm',
+    'actualArm',
+    'consent'
+  ],
   RiskAssessment: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'parent',
@@ -2202,7 +3214,7 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'code',
     'subject',
     'encounter',
-    'occurrence',
+    'occurrence[x]',
     'condition',
     'performer',
     'reasonCode',
@@ -2213,6 +3225,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'note'
   ],
   RiskEvidenceSynthesis: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2246,6 +3266,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'certainty'
   ],
   Schedule: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'active',
     'serviceCategory',
@@ -2256,6 +3284,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'comment'
   ],
   SearchParameter: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -2284,6 +3320,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'component'
   ],
   ServiceRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -2297,11 +3341,11 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'doNotPerform',
     'code',
     'orderDetail',
-    'quantity',
+    'quantity[x]',
     'subject',
     'encounter',
-    'occurrence',
-    'asNeeded',
+    'occurrence[x]',
+    'asNeeded[x]',
     'authoredOn',
     'requester',
     'performerType',
@@ -2319,6 +3363,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'relevantHistory'
   ],
   Slot: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'serviceCategory',
     'serviceType',
@@ -2332,6 +3384,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'comment'
   ],
   Specimen: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'accessionIdentifier',
     'status',
@@ -2346,8 +3406,31 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'condition',
     'note'
   ],
-  SpecimenDefinition: ['identifier', 'typeCollected', 'patientPreparation', 'timeAspect', 'collection', 'typeTested'],
+  SpecimenDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'typeCollected',
+    'patientPreparation',
+    'timeAspect',
+    'collection',
+    'typeTested'
+  ],
   StructureDefinition: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2377,6 +3460,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'differential'
   ],
   StructureMap: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2396,13 +3487,94 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'import',
     'group'
   ],
-  Subscription: ['status', 'contact', 'end', 'reason', 'criteria', 'error', 'channel'],
-  Substance: ['identifier', 'status', 'category', 'code', 'description', 'instance', 'ingredient'],
-  SubstanceNucleicAcid: ['sequenceType', 'numberOfSubunits', 'areaOfHybridisation', 'oligoNucleotideType', 'subunit'],
-  SubstancePolymer: ['class', 'geometry', 'copolymerConnectivity', 'modification', 'monomerSet', 'repeat'],
-  SubstanceProtein: ['sequenceType', 'numberOfSubunits', 'disulfideLinkage', 'subunit'],
-  SubstanceReferenceInformation: ['comment', 'gene', 'geneElement', 'classification', 'target'],
+  Subscription: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'status',
+    'contact',
+    'end',
+    'reason',
+    'criteria',
+    'error',
+    'channel'
+  ],
+  Substance: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'identifier',
+    'status',
+    'category',
+    'code',
+    'description',
+    'instance',
+    'ingredient'
+  ],
+  SubstancePolymer: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'class',
+    'geometry',
+    'copolymerConnectivity',
+    'modification',
+    'monomerSet',
+    'repeat'
+  ],
+  SubstanceProtein: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'sequenceType',
+    'numberOfSubunits',
+    'disulfideLinkage',
+    'subunit'
+  ],
+  SubstanceReferenceInformation: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
+    'comment',
+    'gene',
+    'geneElement',
+    'classification',
+    'target'
+  ],
   SubstanceSourceMaterial: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'sourceMaterialClass',
     'sourceMaterialType',
     'sourceMaterialState',
@@ -2418,6 +3590,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'partDescription'
   ],
   SubstanceSpecification: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'type',
     'status',
@@ -2439,6 +3619,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'sourceMaterial'
   ],
   SupplyDelivery: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'basedOn',
     'partOf',
@@ -2446,20 +3634,28 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'patient',
     'type',
     'suppliedItem',
-    'occurrence',
+    'occurrence[x]',
     'supplier',
     'destination',
     'receiver'
   ],
   SupplyRequest: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'category',
     'priority',
-    'item',
+    'item[x]',
     'quantity',
     'parameter',
-    'occurrence',
+    'occurrence[x]',
     'authoredOn',
     'requester',
     'supplier',
@@ -2469,6 +3665,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'deliverTo'
   ],
   Task: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'instantiatesCanonical',
     'instantiatesUri',
@@ -2502,6 +3706,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'output'
   ],
   TerminologyCapabilities: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'version',
     'name',
@@ -2528,6 +3740,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'closure'
   ],
   TestReport: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'name',
     'status',
@@ -2542,6 +3762,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'teardown'
   ],
   TestScript: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2568,6 +3796,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'teardown'
   ],
   ValueSet: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'url',
     'identifier',
     'version',
@@ -2588,6 +3824,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'expansion'
   ],
   VerificationResult: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'target',
     'targetLocation',
     'need',
@@ -2604,6 +3848,14 @@ export const parsedPropertyPaths: Record<string, string[]> = {
     'validator'
   ],
   VisionPrescription: [
+    'id',
+    'meta',
+    'implicitRules',
+    'language',
+    'text',
+    'contained',
+    'extension',
+    'modifierExtension',
     'identifier',
     'status',
     'created',
