+++ src/scripts/parseProperties.ts
@@ -1,43 +1,53 @@
-/* eslint-disable  @typescript-eslint/no-explicit-any */
+import { ElementDefinition } from 'fhir/r4';
 import * as fs from 'fs';
 import * as path from 'path';
-import * as xml2js from 'xml2js';
 
-const modelInfoPath = path.resolve(path.join(__dirname, '../fhir/fhir-modelinfo-4.0.1.xml'));
-const outputPath = path.resolve(path.join(__dirname, '../data/propertyPaths.ts'));
-const xmlStr = fs.readFileSync(modelInfoPath, 'utf8');
+const STRUCTURE_DEFINITIONS_BASE_PATH = path.join(__dirname, '../fhir/resource-definitions');
+const parsedPropertiesOutputPath = path.resolve(path.join(__dirname, '../data/propertyPaths.ts'));
 
 /**
- * Parse FHIR Model Info XML and output information about what properties exist
- * for each FHIR Resource type.
- * @param {string} xml the string content of the model info XML to parse
- * @return object whose keys are resourceTypes and values correspond to
- * the resourceTypes' properties
+ * Parse the StructureDefinition of resource types supported by this server for properties associated with each type.
+ * The StructureDefinitions of each resource type is found in the FHIR R4 spec
+ * For example, the StructureDefinition for Encounter was found here: https://hl7.org/fhir/R4/encounter.profile.json.html
+ * @returns {Object} object whose keys are resourceTypes and values are arrays of strings that are elements
  */
-export async function parse(xml: string) {
-  const { modelInfo } = await xml2js.parseStringPromise(xml);
-  const domainInfo = modelInfo.typeInfo.filter((ti: any) => ti.$.baseType === 'FHIR.DomainResource');
+async function main() {
+  const files = fs.readdirSync(STRUCTURE_DEFINITIONS_BASE_PATH).map(f => ({
+    shortName: f.split('.profile')[0],
+    fullPath: path.join(STRUCTURE_DEFINITIONS_BASE_PATH, f)
+  }));
 
-  const results: { [key: string]: string[] } = {};
+  const parsedPropertyPathsResults: Record<string, string[]> = {};
 
-  domainInfo.forEach((di: any) => {
-    const resourceType = di.$.name;
-    results[resourceType] = di.element.map((elem: any) => elem.$.name);
+  files.forEach(f => {
+    const properties: string[] = [];
+
+    // read the contents of the file
+    const structureDef = JSON.parse(fs.readFileSync(f.fullPath, 'utf8')) as fhir4.StructureDefinition;
+    structureDef.snapshot?.element.forEach(e => {
+      const elem = e.id?.split('.');
+      if (elem && elem.length === 2) {
+        properties.push(elem[1]);
+      }
+    });
+    if (structureDef.id) {
+      parsedPropertyPathsResults[structureDef.id] = properties;
+    }
   });
-  return results;
+  return parsedPropertyPathsResults;
 }
 
-parse(xmlStr)
-  .then(data => {
+main()
+  .then(parsedPropertyPaths => {
     fs.writeFileSync(
-      outputPath,
+      parsedPropertiesOutputPath,
       `
         export const parsedPropertyPaths: Record<string, string[]> =
-          ${JSON.stringify(data, null, 2)};
+          ${JSON.stringify(parsedPropertyPaths, null, 2)};
         `,
       'utf8'
     );
-    console.log(`Wrote file to ${outputPath}`);
+    console.log(`Wrote file to ${parsedPropertiesOutputPath}`);
   })
   .catch(e => {
     console.error(e);
